# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookStatus = (typeof BookStatus)[keyof typeof BookStatus];

export const BookStatus = {
  currently_reading: "currently_reading",
  completed: "completed",
} as const;

export interface Book {
  id: number;
  title: string;
  author: string;
  totalPages: number;
  currentPage: number;
  status: BookStatus;
  startedAt: string;
  /** @nullable */
  completedAt: string | null;
  createdAt: string;
}

export interface CreateBookRequest {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  /** @minimum 1 */
  totalPages: number;
  /** @minimum 0 */
  currentPage?: number;
}

export interface UpdateBookRequest {
  /** @minLength 1 */
  title?: string;
  /** @minLength 1 */
  author?: string;
  /** @minimum 1 */
  totalPages?: number;
  /** @minimum 0 */
  currentPage?: number;
  status?: BookStatus;
}

export interface CreateReadingSessionRequest {
  bookId: number;
  /** @minimum 1 */
  pagesRead: number;
  /** @minimum 1 */
  minutesRead?: number;
  loggedAt?: string;
}

export interface ReadingSession {
  id: number;
  bookId: number;
  bookTitle: string;
  pagesRead: number;
  /** @nullable */
  minutesRead: number | null;
  loggedAt: string;
}

export interface ReadingGoalRow {
  bookId: number;
  bookName: string;
  pagesCompleted: number;
  totalPages: number;
  estimatedDateCompleted: string;
}

export interface ReadingStats {
  pagesToday: number;
  dailyGoalPages: number;
  activeBooks: number;
  completedBooks: number;
  currentStreakDays: number;
  totalPagesRead: number;
  averagePagesPerDay: number;
}

export interface ReadingActivity {
  id: number;
  bookTitle: string;
  pagesRead: number;
  loggedAt: string;
}

export interface ReadingSettings {
  id: number;
  dailyGoalPages: number;
  reminderTime: string;
  updatedAt: string;
}

export interface UpdateSettingsRequest {
  /** @minimum 1 */
  dailyGoalPages?: number;
  reminderTime?: string;
}