# lib/api-zod/src/generated/types/book.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BookStatus } from "./bookStatus";

export interface Book {
  id: number;
  title: string;
  author: string;
  totalPages: number;
  currentPage: number;
  status: BookStatus;
  startedAt: Date;
  /** @nullable */
  completedAt: Date | null;
  createdAt: Date;
}