# lib/api-zod/src/generated/types/bookStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type BookStatus = (typeof BookStatus)[keyof typeof BookStatus];

export const BookStatus = {
  currently_reading: "currently_reading",
  completed: "completed",
} as const;