# lib/api-zod/src/generated/types/createBookRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookRequest {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  author: string;
  /** @minimum 1 */
  totalPages: number;
  /** @minimum 0 */
  currentPage?: number;
}