# lib/api-zod/src/generated/types/createReadingSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReadingSessionRequest {
  bookId: number;
  /** @minimum 1 */
  pagesRead: number;
  /** @minimum 1 */
  minutesRead?: number;
  loggedAt?: Date;
}