# lib/api-zod/src/generated/types/readingSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReadingSession {
  id: number;
  bookId: number;
  bookTitle: string;
  pagesRead: number;
  /** @nullable */
  minutesRead: number | null;
  loggedAt: Date;
}