# lib/api-zod/src/generated/types/readingSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReadingSettings {
  id: number;
  dailyGoalPages: number;
  reminderTime: string;
  updatedAt: Date;
}