# lib/api-zod/src/generated/types/readingStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReadingStats {
  pagesToday: number;
  dailyGoalPages: number;
  activeBooks: number;
  completedBooks: number;
  currentStreakDays: number;
  totalPagesRead: number;
  averagePagesPerDay: number;
}